refactor: split getPaths into own file
