+++ app/services/flow/pruner.ts
@@ -1,20 +1,13 @@
 import _ from "lodash";
-import type { ArrayData, Context } from "~/flows/contexts";
-import { type FlowId, parsePathname } from "~/flows/flowIds";
-import { flows } from "~/flows/flows.server";
-import { addPageDataToUserData } from "./pageData";
-import {
-  type FlowController,
-  buildFlowController,
-} from "./server/buildFlowController";
-import { type ArrayConfig } from "../array";
+import type { Context } from "~/flows/contexts";
+import type { FlowId } from "~/flows/flowIds";
+import { getPaths, type Path } from "./getPaths";
 import { resolveArrayCharacter } from "../array/resolveArrayCharacter";
 import {
   fetchAllFormFields,
   type FormFieldsMap,
 } from "../cms/fetchAllFormFields";
 
-type Path = { stepIds: string[]; arrayIndex?: number };
 type FormField = { name: string; arrayIndex?: number };
 
 export async function pruneIrrelevantData(
@@ -30,18 +23,6 @@ export async function pruneIrrelevantData(
   return _.pick(userData, propsToKeep);
 }
 
-export function getPaths(userData: Context, flowId: FlowId): Path[] {
-  const flowController = buildFlowController({
-    ...flows[flowId],
-    data: userData,
-  });
-
-  return [
-    getMainFlowPath(flowController),
-    ...getSubflowPaths(flowController, userData),
-  ];
-}
-
 export async function getFormFields(
   paths: Path[],
   formFieldsMap: FormFieldsMap,
@@ -51,60 +32,4 @@ export async function getFormFields(
       (formFieldsMap[`/${stepId}`] ?? []).map((name) => ({ name, arrayIndex })),
     ),
   );
-}
-
-function getSteps(
-  flowController: FlowController,
-  subFlowsInitialStep?: string,
-) {
-  const stepIds: string[] = [];
-  let stepId: string | undefined = parsePathname(
-    subFlowsInitialStep ?? flowController.getInitial(),
-  ).stepId;
-
-  while (stepId) {
-    stepIds.push(stepId);
-    const nextPath = flowController.getNext(stepId);
-    stepId = nextPath ? parsePathname(nextPath).stepId : undefined;
-  }
-
-  return stepIds;
-}
-
-function getMainFlowPath(flowController: FlowController): Path {
-  return { stepIds: getSteps(flowController) };
-}
-
-function getSubflowPaths(
-  flowController: FlowController,
-  userData: Context,
-): Path[] {
-  return Object.entries(flowController.getRootMeta()?.arrays ?? {})
-    .filter(([key]) => !_.isUndefined(userData[key]))
-    .filter(([_key, config]) => userData[config.statementKey] === "yes")
-    .map(([key, arrayConfig]) =>
-      (userData[key] as ArrayData).map((_data, index) => ({
-        stepIds: getSteps(
-          buildFlowController({
-            config: flowController.getConfig(),
-            guards: flowController.getGuards(),
-            data: addPageDataToUserData(userData, {
-              arrayIndexes: [index],
-            }),
-          }),
-          getSubFlowInitialStep(arrayConfig),
-        ).filter(stepBelongsToArray(arrayConfig)),
-        arrayIndex: index,
-      })),
-    )
-    .flat();
-}
-
-function getSubFlowInitialStep(arrayConfig: ArrayConfig): string {
-  return arrayConfig.url + "/" + arrayConfig.initialInputUrl;
-}
-
-function stepBelongsToArray(arrayConfig: ArrayConfig) {
-  return (stepId: string) =>
-    stepId.startsWith(parsePathname(arrayConfig.url).stepId);
 }
